# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  models.Response:
    properties:
      message:
        type: string
    type: object
  models.Site:
    properties:
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.User:
    properties:
      bankAccountName:
        type: string
      bankAccountNumber:
        type: string
      corporateName:
        type: string
      create:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      imageProfile:
        type: string
      lastName:
        type: string
      lastUpdate:
        type: string
      project:
        type: string
      role:
        type: string
      site:
        $ref: '#/definitions/models.Site'
        type: object
      siteId:
        type: string
      slackAccount:
        type: string
      thaiCitizenId:
        type: string
      transcript:
        type: string
      vat:
        type: string
    type: object
  utils.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
host: http://worklog-dev.odds.team/api
info:
  contact: {}
  description: This is a sample server odds-worklog server.
  license: {}
  title: Odds-Worklog Example API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: get user list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: List user
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create User
      parameters:
      - description: id can empty
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Create User
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Delete User
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get User By Id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
            type: object
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Get User By Id
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update User
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: id can empty
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Update User
      tags:
      - users
  /users/site/{id}:
    get:
      consumes:
      - application/json
      description: Get User By Site Id
      parameters:
      - description: Site id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Get User By Site Id
      tags:
      - users
swagger: "2.0"